# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * XEMENS Constructor API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface RegisterInput {
  email: string;
  /** @nullable */
  name?: string | null;
  password: string;
}

export interface RegisterResult {
  ok: boolean;
  message: string;
  pendingEmail: string;
  /** @nullable */
  pendingName?: string | null;
  /** @nullable */
  devCode?: string | null;
}

export interface VerifyInput {
  email: string;
  code: string;
  /** @nullable */
  name?: string | null;
  password: string;
}

export interface VerifyResult {
  ok: boolean;
  message: string;
  userId: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthUser {
  id: string;
  email: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  image?: string | null;
}

export interface AuthSession {
  user: AuthUser;
}

export interface ProjectSummary {
  id: string;
  name: string;
  published: boolean;
  /** @nullable */
  siteId?: string | null;
  updatedAt: string;
  createdAt: string;
}

export interface ProjectList {
  projects: ProjectSummary[];
}

export type ProjectDetailProject = {
  id: string;
  name: string;
  published: boolean;
  /** @nullable */
  siteId?: string | null;
  data: string;
  updatedAt: string;
  createdAt: string;
};

export interface ProjectDetail {
  project: ProjectDetailProject;
}

export interface ProjectInput {
  name?: string;
}

export interface ProjectUpdate {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  data?: string | null;
  /** @nullable */
  description?: string | null;
}

export interface PublishResult {
  ok: boolean;
  siteId: string;
  url: string;
}

export interface SiteRuntime {
  data: string;
  siteId: string;
}