# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: XEMENS Constructor API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: projects
    description: Project management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Send email verification code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "200":
          description: Code sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/verify:
    post:
      operationId: verifyEmail
      tags: [auth]
      summary: Verify email code and create account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyInput"
      responses:
        "200":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login with credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logoutUser
      tags: [auth]
      summary: Logout current user
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current session user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List user's projects
      responses:
        "200":
          description: Projects list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectList"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "200":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  /projects/{id}/publish:
    post:
      operationId: publishProject
      tags: [projects]
      summary: Publish a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /site/{siteId}:
    get:
      operationId: getSiteRuntime
      tags: [projects]
      summary: Get published site data
      parameters:
        - name: siteId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Site data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteRuntime"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok

    RegisterInput:
      type: object
      properties:
        email:
          type: string
        name:
          type: ["string", "null"]
        password:
          type: string
      required:
        - email
        - password

    RegisterResult:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
        pendingEmail:
          type: string
        pendingName:
          type: ["string", "null"]
        devCode:
          type: ["string", "null"]
      required:
        - ok
        - message
        - pendingEmail

    VerifyInput:
      type: object
      properties:
        email:
          type: string
        code:
          type: string
        name:
          type: ["string", "null"]
        password:
          type: string
      required:
        - email
        - code
        - password

    VerifyResult:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
        userId:
          type: string
      required:
        - ok
        - message
        - userId

    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
      required:
        - id
        - email

    AuthSession:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
      required:
        - user

    ProjectSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        published:
          type: boolean
        siteId:
          type: ["string", "null"]
        updatedAt:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - published
        - updatedAt
        - createdAt

    ProjectList:
      type: object
      properties:
        projects:
          type: array
          items:
            $ref: "#/components/schemas/ProjectSummary"
      required:
        - projects

    ProjectDetail:
      type: object
      properties:
        project:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            published:
              type: boolean
            siteId:
              type: ["string", "null"]
            data:
              type: string
            updatedAt:
              type: string
            createdAt:
              type: string
          required:
            - id
            - name
            - published
            - data
            - updatedAt
            - createdAt
      required:
        - project

    ProjectInput:
      type: object
      properties:
        name:
          type: string
      required: []

    ProjectUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        data:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
      required: []

    PublishResult:
      type: object
      properties:
        ok:
          type: boolean
        siteId:
          type: string
        url:
          type: string
      required:
        - ok
        - siteId
        - url

    SiteRuntime:
      type: object
      properties:
        data:
          type: string
        siteId:
          type: string
      required:
        - data
        - siteId