# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * XEMENS Constructor API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthUser {
  id: string;
  email: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  image?: string | null;
}