# lib/api-zod/src/generated/types/projectList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * XEMENS Constructor API
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectSummary } from './projectSummary';

export interface ProjectList {
  projects: ProjectSummary[];
}