# lib/api-zod/src/generated/types/projectSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * XEMENS Constructor API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectSummary {
  id: string;
  name: string;
  published: boolean;
  /** @nullable */
  siteId?: string | null;
  updatedAt: string;
  createdAt: string;
}