# lib/api-zod/src/generated/types/projectUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * XEMENS Constructor API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectUpdate {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  data?: string | null;
  /** @nullable */
  description?: string | null;
}