# lib/api-zod/src/generated/types/registerResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * XEMENS Constructor API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterResult {
  ok: boolean;
  message: string;
  pendingEmail: string;
  /** @nullable */
  pendingName?: string | null;
  /** @nullable */
  devCode?: string | null;
}