# lib/api-zod/src/generated/types/verifyInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * XEMENS Constructor API
 * OpenAPI spec version: 0.1.0
 */

export interface VerifyInput {
  email: string;
  code: string;
  /** @nullable */
  name?: string | null;
  password: string;
}